projects/sgc-reports-mcp/requirements.txt
mcp[cli]>=1.2.0
docxtpl>=0.16.8
python-docx>=1.1.2
xlsxwriter>=3.2.0
markdown>=3.6
weasyprint>=62.3
pydantic>=2.8.2
